chore: move copy_assets into utils
