+++ ui/src/main.rs
@@ -13,7 +13,6 @@ use dioxus_desktop::tao::menu::AboutMetadata;
 use dioxus_desktop::Config;
 use dioxus_desktop::{tao, use_window};
 use extensions::UplinkExtension;
-use filetime::FileTime;
 use futures::channel::oneshot;
 use futures::StreamExt;
 use kit::components::nav::Route as UIRoute;
@@ -24,7 +23,6 @@ use once_cell::sync::Lazy;
 use overlay::{make_config, OverlayDom};
 use rfd::FileDialog;
 use std::collections::{HashMap, HashSet};
-use std::path::{Path, PathBuf};
 use std::time::Instant;
 use std::{fs, io};
 use uuid::Uuid;
@@ -102,98 +100,6 @@ pub enum AuthPages {
     Success(multipass::identity::Identity),
 }
 
-fn copy_assets() {
-    log::debug!("copy_assets");
-    if !STATIC_ARGS.production_mode {
-        return;
-    }
-    let exe_path = match std::env::current_exe() {
-        Ok(p) => p,
-        Err(e) => {
-            log::error!("failed to get path of uplink executable: {e}");
-            return;
-        }
-    };
-
-    let current_version = env!("CARGO_PKG_VERSION");
-    let assets_version_file = STATIC_ARGS.dot_uplink.join("assets_version.txt");
-    let assets_version = std::fs::read_to_string(&assets_version_file).unwrap_or_default();
-    if current_version == assets_version {
-        let exe_meta =
-            fs::metadata(&exe_path).expect("failed to get metadata for uplink executable");
-        let version_meta =
-            fs::metadata(&assets_version_file).expect("failed to get metadata for assets version");
-        let exe_changed = FileTime::from_last_modification_time(&exe_meta);
-        let assets_changed = FileTime::from_last_modification_time(&version_meta);
-        if assets_changed > exe_changed {
-            log::debug!("assets already exist");
-            return;
-        } else {
-            log::debug!("re-install suspected. copying over assets");
-        }
-    }
-
-    let assets_path = if cfg!(target_os = "windows") {
-        match exe_path
-            .parent()
-            .and_then(|x| x.parent())
-            .map(|x| x.join("extra.zip"))
-        {
-            Some(p) => p,
-            None => {
-                log::error!("failed to get parent directory of uplink executable");
-                return;
-            }
-        }
-    } else if cfg!(any(
-        target_os = "linux",
-        target_os = "macos",
-        target_os = "ios"
-    )) {
-        PathBuf::from("/opt/satellite-im/uplink/extra.zip")
-    } else {
-        log::error!("unknown OS type. failed to copy assets");
-        return;
-    };
-
-    if let Err(e) = std::fs::remove_dir_all(&STATIC_ARGS.extras_path) {
-        log::error!("failed to delete old assets directory: {e}");
-    }
-    if let Err(e) = unzip_archive(&assets_path, &STATIC_ARGS.extras_path) {
-        log::error!("failed to unizp assets archive {assets_path:?}: {e}");
-    }
-
-    if let Err(e) = std::fs::write(assets_version_file, current_version) {
-        log::error!("failed to save assets_version_file: {e}");
-    }
-}
-
-// taken from https://github.com/zip-rs/zip/blob/master/examples/extract.rs
-fn unzip_archive(src: &Path, dest: &Path) -> Result<(), Box<dyn std::error::Error>> {
-    let assets_zip = fs::File::open(src)?;
-    let mut archive = zip::ZipArchive::new(assets_zip)?;
-    for idx in 0..archive.len() {
-        let mut file = archive.by_index(idx)?;
-        let outpath = match file.enclosed_name() {
-            Some(path) => dest.join(path),
-            None => continue,
-        };
-        if (*file.name()).ends_with('/') || (*file.name()).ends_with('\\') {
-            fs::create_dir_all(&outpath)?;
-        } else {
-            if let Some(p) = outpath.parent() {
-                if !p.exists() {
-                    fs::create_dir_all(p)?;
-                }
-            }
-            let mut outfile = fs::File::create(&outpath)?;
-            io::copy(&mut file, &mut outfile)?;
-        }
-    }
-
-    Ok(())
-}
-
 fn main() {
     // Attempts to increase the file desc limit on unix-like systems
     // Note: Will be changed out in the future
@@ -264,7 +170,7 @@ fn main() {
     std::fs::create_dir_all(&STATIC_ARGS.themes_path).expect("error creating themes directory");
     std::fs::create_dir_all(&STATIC_ARGS.fonts_path).expect("error fonts themes directory");
 
-    copy_assets();
+    utils::copy_assets();
 
     let mut main_menu = Menu::new();
     let mut app_menu = Menu::new();
